moved all css into dedicated assets folder
