+++ todo-list/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import ListReducer from "./Reducer";
 import NewList from "./New"
 import CurrentList from "./Current"
-import "./App.css"
+import "./Assets/css/App.css"
 
 const App = () => {
   const [todoList, dispatchTodoList] = useReducer(ListReducer,
